Internship_project/requirements.txt
Django
python-dateutil
requests
requests-oauthlib
